main.py changes
Changed main.py name to game.py, created main.py to run game.
